requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai
google-generativeai